feat(service): improve update scheduling logic using historical logs and retry backoff
- Calculate lowest, average, and highest update delays from historical log data to determine the next update time.
- Enforce a minimum 30-minute interval between successful updates.
- Implement a retry strategy for failed updates using exponential backoff (up to 30 minutes) and random jitter.
- Refactor `lowestDelay` to use `Long` for consistency across domain and DBO layers.

Fixes #5

+++ src/main/kotlin/net/jonasmf/auctionengine/dbo/dynamodb/AuctionHouseDynamo.kt
@@ -42,7 +42,7 @@ data class AuctionHouseDynamo(
     var lastStatsInsert: Instant? = null,
     @get:DynamoDbConvertedBy(KotlinInstantAsLongAttributeConverter::class)
     var lastTrendUpdateInitiation: Instant? = null,
-    var lowestDelay: Int = 0,
+    var lowestDelay: Long = 0,
     @get:DynamoDbSecondarySortKey(indexNames = ["region-index"])
     @get:DynamoDbConvertedBy(KotlinInstantAsLongAttributeConverter::class)
     var nextUpdate: Instant? = null,
